projects/09-control-coverage-mapper/requirements.txt
pytest>=8.0.0
python-docx>=0.8.11
pdfplumber>=0.7.0
